Change safety requirements of DMA traits
As @adamgreig mentioned in #1, it is not possible for implementing types to
guarantee that a memory region will be readable by DMA, if we don't want to
restrict the DMA buffer types too much.

Instead the traits should only ensure that the buffer cannot be freed as long as
there might still be a DMA transfer accessing it. Checking if the buffer is in a
memory region that is accessible by DMA needs to be done at runtime instead.